requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter == 23.1
drf-spectacular>=0.15.1,<0.16 
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
